StatusBar/QS: Redesign - Part 1
Its here, finally. Still not done yet, might look weird currently but its just the basis still, I will
push another commit for this again tomorrow in time for our test build that is soon :P Anyways, currently
this commit puts all the images in for QS and Statusbar - it changes most of the textcolors hopefully
and has some QS tile background changes. It might be more but this has been on my local for around a month and
I forget what I have done so far lol but yes, thats why this is incomplete still. More commits to come :)

Thanks to Dean Jordan for all graphics!

Bye bye

Signed-off-by: Chintan Shah <cgintan1998@gmail.com>